move reading data to dataset object
